Refactor imports in content and search routes to streamline code structure and improve readability
